changed point type to point2d from point2f

+++ industrial_calibration_libs/include/industrial_calibration_libs/observations.h
@@ -28,7 +28,7 @@ struct Observation
 };
 #endif
 
-typedef std::vector<cv::Point2f> ObservationPoints;
+typedef std::vector<cv::Point2d> ObservationPoints;
 
 // Note(gChiou): I think the above typedef is for a single image.
 // The input of the ObservationExtractor class should be a vector of
